feat(bench): paper-aligned Edge/Base/Pro bench scripts (Qwen3 256d + gte-reranker-modernbert)
Phase 2 of v0.4.0 paper-aligned realignment.

- Renamed bench_truememory_base.py → bench_truememory_edge.py
  (paper §2.0 Edge tier: Model2Vec potion-base-8M @ 256d +
  cross-encoder/ms-marco-MiniLM-L-6-v2, HyDE OFF, CPU-only, target 90.1%).
- bench_truememory_pro.py now uses set_embedding_model("pro") and
  Alibaba-NLP/gte-reranker-modernbert-base. HyDE ON. Target 91.8%
  (paper §2.0 +HyDE).
- New bench_truememory_base.py (paper §2.0 Default): same embedder +
  reranker as Pro, HyDE OFF. Target 91.5%.
- File-level ruff noqa (E701, E702, E722, F541, F841) added to all three
  bench scripts to permit the intentionally terse one-line-per-statement
  Modal-ship style without flagging pre-existing style choices.

Note on git rename detection: because a new file is created at the old
bench_truememory_base.py path in the same commit, git's default diff
reports "M base.py" + "A edge.py"; `git diff --find-copies-harder` shows
the intent clearly as C091 (91%-similar copy from old base → edge). The
renamed content is bench_truememory_edge.py; the new Base file is a
different script (Qwen3 256d Default config).

Spec: _working/PHASED_REALIGNMENT_PLAN.md §Phase 2
