+++ benchmarks/locomo/scripts/bench_truememory_edge.py
@@ -1,22 +1,21 @@
 #!/usr/bin/env python3
 """
-LoCoMo Benchmark — TrueMemory Pro Tier (+HyDE, T4 GPU)
-======================================================
-TrueMemory Pro tier using Qwen3-Embedding-0.6B @ 256d Matryoshka embeddings,
-gte-reranker-modernbert-base (149M params) reranker, and HyDE (Hypothetical
-Document Embeddings) via OpenRouter LLM. Paper §2.0 target: 91.8%.
-Requires a T4 GPU on Modal.
+LoCoMo Benchmark — TrueMemory Edge Tier
+======================================
+TrueMemory Edge tier using Model2Vec (potion-base-8M, 256-dim) embeddings
+and cross-encoder reranker (ms-marco-MiniLM-L-6-v2, 22M params).
+No HyDE, no GPU required. CPU-only, ~30M total params. Paper §2.0 target: 90.1%.
 
 This is a fully self-contained Modal script. No local imports required.
 
-Dependencies: truememory[gpu], sentence-transformers
+Dependencies: truememory, sentence-transformers
 Eval: openai/gpt-4.1-mini (answers) + openai/gpt-4o-mini (judge) via OpenRouter
 
 Usage:
     modal secret create openrouter-key OPENROUTER_API_KEY=sk-or-...
 
-    modal run --detach bench_truememory_pro.py          # Full run (10 convs, 1540 Qs)
-    modal run --detach bench_truememory_pro.py --smoke  # Smoke test (1 conv, 5 Qs)
+    modal run --detach bench_truememory_edge.py          # Full run (10 convs, 1540 Qs)
+    modal run --detach bench_truememory_edge.py --smoke  # Smoke test (1 conv, 5 Qs)
 
     modal volume get locomo-results / ./results --force
 """
@@ -36,13 +35,13 @@ from pathlib import Path
 
 # ── Modal Setup ──────────────────────────────────────────────────────────
 
-app = modal.App("locomo-truememory-pro")
+app = modal.App("locomo-truememory-edge")
 vol = modal.Volume.from_name("locomo-results", create_if_missing=True)
 VM = "/results"
 OPENROUTER_BASE_URL = "https://openrouter.ai/api/v1"
 
 img = (modal.Image.debian_slim(python_version="3.11")
-    .pip_install("openai>=1.0", "truememory[gpu]", "sentence-transformers"))
+    .pip_install("openai>=1.0", "truememory", "sentence-transformers"))
 
 # ── Eval Config (IDENTICAL across all systems) ──────────────────────────
 
@@ -202,28 +201,17 @@ def _nm_format_ctx(results):
         parts.append(f"{meta} {r['content']}")
     return "\n\n".join(parts)
 
-def _nm_make_hyde_fn():
-    """Create HyDE LLM callable via OpenRouter — matches v2 scripts."""
-    client = mkc()
-    def _call(prompt):
-        resp = client.chat.completions.create(
-            model="openai/gpt-4.1-mini", max_tokens=300, temperature=0.3,
-            messages=[{"role": "user", "content": prompt}])
-        return resp.choices[0].message.content
-    return _call
+# ── TrueMemory Edge Retrieval ─────────────────────────────────────────────
 
-# ── TrueMemory Pro Retrieval ──────────────────────────────────────────────
-
-def retrieve_truememory_pro(conv_data, conv_idx):
+def retrieve_truememory_edge(conv_data, conv_idx):
     from truememory.vector_search import set_embedding_model
-    set_embedding_model("pro")
+    set_embedding_model("edge")
     from truememory.engine import TrueMemoryEngine
     from truememory.reranker import get_reranker
     import tempfile
-    get_reranker(model_name="Alibaba-NLP/gte-reranker-modernbert-base")
-    llm_fn = _nm_make_hyde_fn()
+    get_reranker(model_name="cross-encoder/ms-marco-MiniLM-L-6-v2")
     msgs = parse_conv(conv_data)
-    tmp_db = tempfile.mktemp(suffix=".db", prefix=f"pro_{conv_idx}_")
+    tmp_db = tempfile.mktemp(suffix=".db", prefix=f"edge_{conv_idx}_")
     engine = TrueMemoryEngine(db_path=tmp_db)
     # Ingest
     import json as _json
@@ -233,11 +221,10 @@ def retrieve_truememory_pro(conv_data, conv_idx):
                  for m in msgs]
     with open(tmp_json,"w") as f: _json.dump(msg_dicts, f)
     engine.ingest(tmp_json)
-    # Retrieve with HyDE + reranker
+    # Retrieve
     results = []
     for qa in get_qa(conv_data):
-        sr = engine.search_agentic(qa["question"], limit=100, llm_fn=llm_fn,
-                                   use_hyde=True, use_reranker=True)
+        sr = engine.search_agentic(qa["question"], limit=100, use_hyde=False, use_reranker=True)
         ctx = _nm_format_ctx(sr)
         results.append((qa["question"], qa["category"], qa["answer"], ctx or "No results found."))
     engine.close()
@@ -256,12 +243,12 @@ def _bench_conv(conv_data, conv_idx, smoke=False):
     qas = get_qa(conv_data)
     n_qs = 5 if smoke else len(qas)
 
-    print(f"  [truememory_pro] Conv {conv_idx} ({sid}): {len(parse_conv(conv_data))} msgs, {n_qs} Qs")
+    print(f"  [truememory_edge] Conv {conv_idx} ({sid}): {len(parse_conv(conv_data))} msgs, {n_qs} Qs")
 
     # Step 1: Retrieve
     t0 = time.time()
     try:
-        all_retrieved = retrieve_truememory_pro(conv_data, conv_idx)
+        all_retrieved = retrieve_truememory_edge(conv_data, conv_idx)
     except Exception as e:
         print(f"    RETRIEVAL FAILED: {e}")
         return [{"question":q["question"],"category":q["category"],"gold_answer":q["answer"],
@@ -298,7 +285,7 @@ def _bench_conv(conv_data, conv_idx, smoke=False):
     return details
 
 @app.function(image=img, secrets=[modal.Secret.from_name("openrouter-key")],
-              timeout=14400, memory=8192, gpu="T4")
+              timeout=14400, memory=8192)
 def worker(conv_data, conv_idx, smoke=False):
     return _bench_conv(conv_data, conv_idx, smoke)
 
@@ -307,10 +294,10 @@ def worker(conv_data, conv_idx, smoke=False):
 @app.function(image=img, secrets=[modal.Secret.from_name("openrouter-key")],
               timeout=28800, memory=2048, volumes={VM: vol})
 def orchestrate(locomo_data: list, smoke: bool = False):
-    """Run TrueMemory Pro, checkpointing after each conversation."""
-    system = "truememory_pro"
+    """Run TrueMemory Edge, checkpointing after each conversation."""
+    system = "truememory_edge"
     ckpt_path = f"{VM}/{system}_checkpoint.json"
-    result_path = f"{VM}/{system}_v3_modal.json"
+    result_path = f"{VM}/{system}_v2_run1.json"
     n_convs = 1 if smoke else len(locomo_data)
     mode = "SMOKE TEST" if smoke else "FULL RUN"
 
@@ -378,7 +365,7 @@ def orchestrate(locomo_data: list, smoke: bool = False):
 
     # Save final result
     result = {
-        "system": system, "version": "v3-modal-rerun", "run": 1,
+        "system": system, "version": "v3-checkpointed", "run": 1,
         "answer_model": ANSWER_MODEL, "answer_max_tokens": ANSWER_MAX_TOKENS,
         "answer_temperature": ANSWER_TEMPERATURE, "judge_model": JUDGE_MODEL,
         "judge_max_tokens": JUDGE_MAX_TOKENS, "judge_temperature": JUDGE_TEMPERATURE,
@@ -409,7 +396,7 @@ def main(smoke: bool = False, dataset: str = None):
         data = json.load(f)
 
     print(f"\n{'='*60}")
-    print(f"LoCoMo Benchmark — TrueMemory Pro (+HyDE, T4 GPU) — {'SMOKE TEST' if smoke else 'FULL RUN'}")
+    print(f"LoCoMo Benchmark — TrueMemory Edge (CPU) — {'SMOKE TEST' if smoke else 'FULL RUN'}")
     print(f"{'='*60}")
     print(f"  Mode:     {'1 conv x 5 Qs' if smoke else '10 convs x 1540 Qs'}")
     print(f"  Answer:   {ANSWER_MODEL} via OpenRouter")
